working for positive changes in the directory

+++ file-download.py
@@ -2,8 +2,8 @@
 This script periodically checks for any updates in the dropbox folder
 and downloads any changes that have occured. It can be run in the 
 background manually or specified as a cron job.
-The variable LOCAL_DIRECTORY can be set up in init.py to change the
-local directory to be synchronized.
+The variable LOCAL_UP_DIRECTORY and LOCAL_DOWN_DIRECTORY can be set
+up in init.py to change the local directory to be synchronized.
 Other variables can also be initialized using init.py
 """
 
@@ -11,7 +11,7 @@ import time
 import sys
 import dropbox
 import os
-from init import ACCESS_TOKEN, DROPBOX_ROOT_DIRECTORY
+from init import ACCESS_TOKEN, DROPBOX_ROOT_DIRECTORY, LOCAL_UP_DIRECTORY, LOCAL_DOWN_DIRECTORY
 import unicodedata
 
 #Starting a connection
@@ -19,36 +19,43 @@ client = dropbox.Dropbox(ACCESS_TOKEN)
 
 #Getting the contents of the dropbox folder
 dropbox_results = client.files_list_folder("",recursive=True)
-
 dropbox_results = dropbox_results.entries
 
 #iterating over the contents of the dropbox folder
 for item in dropbox_results:
 	if isinstance(item,dropbox.files.FileMetadata):
+		
 		#getting file paths
 		file_path = item.path_display
-		print file_path
-		print sys.getsizeof(file_path)
 
-		#file_path is returned in unicode and only to save space we
-		#can convert it to ascii
+		#file_path is returned in unicode and to save space we can
+		#convert it to ascii
 		file_path=unicodedata.normalize('NFKD', file_path).encode('ascii','ignore')
-		print sys.getsizeof(file_path)
+		
+		#downloading the file
 		metadata, response = client.files_download(file_path)
 		file_data = response.content
 		
-		#setting the local path for the file
+		#setting the local path for the file using LOCAL_DOWN_DIRECTORY
 		file_path = file_path.replace(DROPBOX_ROOT_DIRECTORY,"/home/",1)
-		print file_path
-		print os.path.dirname(file_path)
+		file_path = file_path.replace(LOCAL_UP_DIRECTORY,LOCAL_DOWN_DIRECTORY)
+
+		#checking to see if the file already exists or not. If it does
+		#we'll remove the old inconsistent data from it and write new
+		#data into it.
+		#Else, we'll simply create a new file
 		if os.path.exists(file_path):
-			with open(file_path,'rb') as f:
-				data2 = f.read()
-				if file_data!=data2:
-					print "data is not same so gonna heckin' bamboozlin' deletin and rewritin'"
+			with open(file_path,'rw') as f:
+				pre_data = f.read()
+				if file_data!=pre_data:
 					f.seek(0)
 					f.truncate()
 					f.write(file_data)
 		else:
-			with open(file_path, "w+") as f:
-				f.write(file_data)
+			if os.path.exists(os.path.dirname(file_path)):
+				with open(file_path, "w+") as f:
+					f.write(file_data)
+			else:
+				os.mkdir(os.path.dirname(file_path),0755)
+				with open(file_path, "w+") as f:
+					f.write(file_data)
